requirements.txt
asgiref==3.5.0
Django==4.0.3
django-shortuuidfield==0.1.3
Pillow==9.1.0
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
